Add an option to jitter shadow map dithering pattern across frames
If set to Auto (the default), jittering is enabled whenever TAA
is also enabled. With TAA, jittering will look the best regardless
of the monitor refresh rate, framerate or monitor response time.

If set to Always, when TAA is disabled, this makes the dithering pattern
less noticeable as the monitor will smear across frames (due to its
reponse time not being instant). When TAA is disabled, the effect
works best on high refresh rate monitors (> 80 Hz) and
at high framerates (> 80 FPS). That said, even at lower framerates,
jittering the shadow pattern may still be desired to prevent the
dithering pattern from "following" the camera as it moves and rotates.

If set to Never, the existing behavior before this commit is used.
The shadow filtering quality won't improve when TAA is enabled.